requirements.txt
flask-restplus
pynamodb
six
